+++ Source/Core/Core/HW/GCPad.h
@@ -18,8 +18,8 @@ void LoadConfig();
 
 InputConfig* GetConfig();
 
-GCPadStatus GetStatus(u8 pad_num);
-void Rumble(u8 pad_num, ControlState strength);
+GCPadStatus GetStatus(int pad_num);
+void Rumble(int pad_num, ControlState strength);
 
-bool GetMicButton(u8 pad_num);
+bool GetMicButton(int pad_num);
 }
